Update map_temperature.py
re-ran for 2020-21
turns out the need for replace_duplicate_high was that I didn't like the color I chose for 100 (hah!), coupled with the fact that the colorbar was choosing colors from low to high, skipping the second-to-highest color and then showing the highest color.
Fixed this by dropping the highest color (turns out I was duplicating my work from when I rounded up to grab all the necessary bins). Doing this removed the need for the keyword argument replace_dublicate_high (so I removed this) but didn't mess up remove_duplicate_high (I did edit this so it would still work) -the previous plots are still correct this change just removes the hardcoding that I had to do for that and makes the code work for non-duplicate high situations